+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/screen/update/view/UpdateView.kt
@@ -7,22 +7,27 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxHeight
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.wrapContentHeight
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Update
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.verticalScroll
 import androidx.compose.material3.CircularProgressIndicator
+import androidx.compose.material3.DividerDefaults
+import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.SnackbarHost
 import androidx.compose.material3.SnackbarHostState
+import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.res.vectorResource
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import co.electriccoin.zcash.ui.R
 import co.electriccoin.zcash.ui.design.component.Body
@@ -30,7 +35,6 @@ import co.electriccoin.zcash.ui.design.component.GradientSurface
 import co.electriccoin.zcash.ui.design.component.PrimaryButton
 import co.electriccoin.zcash.ui.design.component.Reference
 import co.electriccoin.zcash.ui.design.component.SmallTopAppBar
-import co.electriccoin.zcash.ui.design.component.TertiaryButton
 import co.electriccoin.zcash.ui.design.theme.ZcashTheme
 import co.electriccoin.zcash.ui.fixture.UpdateInfoFixture
 import co.electriccoin.zcash.ui.screen.update.UpdateTag
@@ -73,28 +77,21 @@ fun Update(
                 updateInfo,
                 onDownload,
                 onLater,
-                modifier =
-                    Modifier
-                        .fillMaxWidth()
-                        .padding(
-                            top = ZcashTheme.dimens.spacingDefault,
-                            bottom = ZcashTheme.dimens.spacingHuge,
-                            start = ZcashTheme.dimens.screenHorizontalSpacingBig,
-                            end = ZcashTheme.dimens.screenHorizontalSpacingBig
-                        )
+                modifier = Modifier.fillMaxWidth()
             )
         }
     ) { paddingValues ->
-        UpdateContentNormal(
-            onReference,
+        UpdateContentContent(
+            onReference = onReference,
+            updateInfo = updateInfo,
             modifier =
                 Modifier
                     .fillMaxWidth()
                     .padding(
                         top = paddingValues.calculateTopPadding(),
                         bottom = paddingValues.calculateBottomPadding(),
-                        start = ZcashTheme.dimens.spacingDefault,
-                        end = ZcashTheme.dimens.spacingDefault
+                        start = ZcashTheme.dimens.screenHorizontalSpacingRegular,
+                        end = ZcashTheme.dimens.screenHorizontalSpacingRegular
                     )
         )
     }
@@ -136,6 +133,7 @@ private fun UpdateTopAppBar(updateInfo: UpdateInfo) {
 }
 
 @Composable
+@Suppress("LongMethod")
 private fun UpdateBottomAppBar(
     updateInfo: UpdateInfo,
     onDownload: (state: UpdateState) -> Unit,
@@ -146,73 +144,143 @@ private fun UpdateBottomAppBar(
         modifier = modifier,
         horizontalAlignment = Alignment.CenterHorizontally
     ) {
-        PrimaryButton(
-            onClick = { onDownload(UpdateState.Running) },
-            text = stringResource(R.string.update_download_button),
+        HorizontalDivider(
+            thickness = DividerDefaults.Thickness,
+            color = ZcashTheme.colors.dividerColor
+        )
+
+        Column(
             modifier =
                 Modifier
-                    .testTag(UpdateTag.BTN_DOWNLOAD)
-                    .fillMaxWidth(),
-            enabled = updateInfo.state != UpdateState.Running,
-            outerPaddingValues = PaddingValues(all = ZcashTheme.dimens.spacingNone),
-        )
+                    .padding(
+                        top = ZcashTheme.dimens.spacingDefault,
+                        bottom = ZcashTheme.dimens.spacingBig,
+                        start = ZcashTheme.dimens.screenHorizontalSpacingBig,
+                        end = ZcashTheme.dimens.screenHorizontalSpacingBig
+                    ),
+            horizontalAlignment = Alignment.CenterHorizontally
+        ) {
+            PrimaryButton(
+                onClick = { onDownload(UpdateState.Running) },
+                text = stringResource(R.string.update_download_button),
+                modifier =
+                    Modifier
+                        .testTag(UpdateTag.BTN_DOWNLOAD)
+                        .fillMaxWidth(),
+                enabled = updateInfo.state != UpdateState.Running,
+                outerPaddingValues = PaddingValues(all = ZcashTheme.dimens.spacingNone),
+            )
 
-        Spacer(modifier = Modifier.height(ZcashTheme.dimens.spacingDefault))
+            Spacer(modifier = Modifier.height(ZcashTheme.dimens.spacingDefault))
 
-        TertiaryButton(
-            onClick = onLater,
-            text =
-                stringResource(
-                    updateInfo.isForce.let { force ->
-                        if (force) {
-                            R.string.update_later_disabled_button
+            if (updateInfo.isForce) {
+                Text(
+                    text = stringResource(R.string.update_later_disabled_button),
+                    textAlign = TextAlign.Center,
+                    style = ZcashTheme.typography.primary.bodyLarge,
+                    fontWeight = FontWeight.SemiBold,
+                    modifier =
+                        Modifier
+                            .padding(all = ZcashTheme.dimens.spacingDefault)
+                            .testTag(UpdateTag.BTN_LATER)
+                )
+            } else {
+                Reference(
+                    text = stringResource(R.string.update_later_enabled_button),
+                    onClick = {
+                        if (updateInfo.state != UpdateState.Running) {
+                            onLater()
                         } else {
-                            R.string.update_later_enabled_button
+                            // Keep current state
                         }
-                    }
-                ),
-            modifier = Modifier.testTag(UpdateTag.BTN_LATER),
-            enabled = !updateInfo.isForce && updateInfo.state != UpdateState.Running,
-            outerPaddingValues = PaddingValues(top = ZcashTheme.dimens.spacingSmall)
-        )
+                    },
+                    textAlign = TextAlign.Center,
+                    modifier =
+                        Modifier
+                            .padding(all = ZcashTheme.dimens.spacingDefault)
+                            .testTag(UpdateTag.BTN_LATER)
+                )
+            }
+        }
     }
 }
 
 @Composable
-private fun UpdateContentNormal(
+@Suppress("LongMethod")
+private fun UpdateContentContent(
     onReference: () -> Unit,
-    modifier: Modifier = Modifier
+    updateInfo: UpdateInfo,
+    modifier: Modifier = Modifier,
 ) {
+    val appName = stringResource(id = R.string.app_name)
+
     Column(
-        modifier = modifier,
+        modifier =
+            modifier.then(
+                Modifier
+                    .fillMaxHeight()
+                    .verticalScroll(
+                        rememberScrollState()
+                    )
+            ),
         horizontalAlignment = Alignment.CenterHorizontally
     ) {
-        // Replace this placeholder graphic once this screen is being redesigned
-        @Suppress("MagicNumber")
+        Spacer(modifier = Modifier.height(ZcashTheme.dimens.spacingBig))
+
         Image(
-            imageVector = Icons.Filled.Update,
-            contentDescription = stringResource(id = R.string.update_image_content_description),
-            modifier = Modifier.fillMaxSize(0.45f)
+            imageVector =
+                if (updateInfo.isForce) {
+                    ImageVector.vectorResource(R.drawable.ic_zashi_logo_update_required)
+                } else {
+                    ImageVector.vectorResource(R.drawable.ic_zashi_logo_update_available)
+                },
+            contentDescription = stringResource(id = R.string.update_image_content_description)
         )
 
-        Body(
-            text = stringResource(id = R.string.update_description),
-            modifier =
-                Modifier
-                    .wrapContentHeight()
-                    .align(Alignment.CenterHorizontally)
+        Spacer(modifier = Modifier.height(ZcashTheme.dimens.spacingBig))
+
+        Text(
+            text =
+                if (updateInfo.isForce) {
+                    stringResource(id = R.string.update_title_required)
+                } else {
+                    stringResource(id = R.string.update_title_available, appName)
+                },
+            style = ZcashTheme.extendedTypography.updateTitleStyle,
+            textAlign = TextAlign.Center
         )
+
+        Spacer(modifier = Modifier.height(ZcashTheme.dimens.spacingLarge))
+
+        Body(
+            text =
+                if (updateInfo.isForce) {
+                    stringResource(id = R.string.update_description_required, appName)
+                } else {
+                    stringResource(id = R.string.update_description_available, appName)
+                },
+            textAlign = TextAlign.Center,
+            color = ZcashTheme.colors.textDescriptionDark
+        )
+
+        Spacer(modifier = Modifier.height(ZcashTheme.dimens.spacingDefault))
 
         Reference(
             text = stringResource(id = R.string.update_link_text),
             onClick = {
-                onReference()
+                if (updateInfo.state != UpdateState.Running) {
+                    onReference()
+                } else {
+                    // Keep current state
+                }
             },
-            modifier =
-                Modifier
-                    .wrapContentHeight()
-                    .align(Alignment.CenterHorizontally)
-                    .padding(all = ZcashTheme.dimens.spacingDefault),
+            fontWeight = FontWeight.Normal,
+            textStyle = ZcashTheme.typography.primary.bodyMedium,
+            textAlign = TextAlign.Center,
+            color = ZcashTheme.colors.textDescriptionDark,
+            modifier = Modifier.padding(all = ZcashTheme.dimens.spacingDefault)
         )
+
+        Spacer(modifier = Modifier.height(ZcashTheme.dimens.spacingDefault))
     }
 }